SDIT-405: ✨ Switch to new prisoner numbers endpoint and cp dev instance (#35)

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/prisonersearchindexer/config/IndexBuildProperties.kt
@@ -3,4 +3,4 @@ package uk.gov.justice.digital.hmpps.prisonersearchindexer.config
 import org.springframework.boot.context.properties.ConfigurationProperties
 
 @ConfigurationProperties(prefix = "index-build")
-data class IndexBuildProperties(val pageSize: Long, val completeThreshold: Long)
+data class IndexBuildProperties(val pageSize: Int, val completeThreshold: Long)
